refactor: use the common module to share more code
A lot of the code was still duplicated between the two modules, move more of the shared logic to
being defined in the BaseMultipleSpaceDetector, with the platform specific ones only needing to
provide the report method. This will still work correctly as each platform has to provide its own
Issue with its own Implementation, so they will still be separated
